ENG-4410 add the bits and pieces missing from the last commit

+++ src/ee/common/ValueFactory.hpp
@@ -99,7 +99,7 @@ public:
         return NValue::getAllocatedArrayValueFromString(txt);
     }
 
-    static NValue getArrayValueFromSizeAndType(int elementCount, ValueType elementType)
+    static NValue getArrayValueFromSizeAndType(size_t elementCount, ValueType elementType)
     {
         return NValue::getAllocatedArrayValueFromSizeAndType(elementCount, elementType);
     }
